Begin migrating to events v2 API
This involves gutting some things, particularly service aliases. If you
already had service aliases, make sure to delete the hash from Redis.

The way this should work going forward is that you can search for
services with an events V2 integration key, and those can be aliased to
a particular short name that call can then publish an event for. It's
kind of icky but it helps with PagerDuty's kind of not too friendly
APIs when it comes to building tools (not terrible, just leaves
something to be desired).